[PIR #907] Extract ResolvedEnrichmentCache into its own module + test file
